Rename folder name to conform to standards
