+++ lib/tic_tac_toe_core/game_types.rb
