+++ lib/tic_tac_toe_core/marker.rb
