+++ lib/tic_tac_toe_core/player/computer_player.rb
