+++ lib/tic_tac_toe_core/player/human_player.rb
