+++ lib/tic_tac_toe_core/version.rb
