+++ lib/tic_tac_toe_core/win_case_calculator.rb
