+++ spec/tic_tac_toe_core/board_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
-require 'tictactoe_core/board'
-require 'tictactoe_core/board_helper'
+require 'tic_tac_toe_core/board'
+require 'tic_tac_toe_core/board_helper'
 
 module TicTacToeCore
   describe Board do
