+++ spec/tic_tac_toe_core/fakes/game.rb
