+++ spec/tic_tac_toe_core/fakes/ui_mock.rb
